+++ CHANGELOG.md
@@ -5,7 +5,13 @@ Note: Prior to version 0.3.4.0, this library was named
 `small-bytearray-builder` is now just a compatibility shim
 to ease the migration process.
 
-## 0.3.8.0 -- 2020-??-??
+## 0.3.9.0 -- 2021-11-19
+
+* Add `cstring#`
+* Add `ToBuilder` and `ToBoundedBuilder` classes
+* Add quasiquoter named `bldr` in `Data.Bytes.Builder.Template`.
+
+## 0.3.8.0 -- 2021-06-25
 
 * Fix `doubleDec`, which was encoding small numbers incorrectly.
 * Add `runByteString` for producing `ByteString` from bounded builders.